api/requirements.txt
aniso8601==9.0.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
decorator==5.1.1
Django==4.2.3
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
executing==1.2.0
idna==3.4
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.39
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
yarg==0.1.9
